add @register_model and @register_callback decorators
Refactor ClockTrainer to include Callback
