+++ src/refiners/training_utils/__init__.py
@@ -4,8 +4,20 @@ from importlib.metadata import requires
 
 from packaging.requirements import Requirement
 
-from refiners.training_utils.config import BaseConfig
-from refiners.training_utils.trainer import Trainer
+from refiners.training_utils.callback import Callback, CallbackConfig
+from refiners.training_utils.clock import ClockConfig
+from refiners.training_utils.config import (
+    BaseConfig,
+    ModelConfig,
+    OptimizerConfig,
+    Optimizers,
+    SchedulerConfig,
+    SchedulerType,
+    TrainingConfig,
+)
+from refiners.training_utils.gradient_clipping import GradientClippingConfig
+from refiners.training_utils.trainer import Trainer, register_callback, register_model
+from refiners.training_utils.wandb import WandbConfig, WandbMixin
 
 refiners_requires = requires("refiners")
 assert refiners_requires is not None
@@ -29,4 +41,18 @@ for dep in refiners_requires:
 __all__ = [
     "Trainer",
     "BaseConfig",
+    "ModelConfig",
+    "register_callback",
+    "register_model",
+    "Callback",
+    "CallbackConfig",
+    "WandbMixin",
+    "WandbConfig",
+    "SchedulerConfig",
+    "OptimizerConfig",
+    "TrainingConfig",
+    "ClockConfig",
+    "GradientClippingConfig",
+    "Optimizers",
+    "SchedulerType",
 ]
