+++ src/refiners/training_utils/trainer.py
@@ -1,15 +1,12 @@
-import random
-import time
 from abc import ABC, abstractmethod, abstractproperty
+from dataclasses import dataclass
 from functools import cached_property, wraps
-from typing import Any, Callable, Generic, Iterable, TypeVar, cast
+from typing import Any, Callable, Generic, Literal, TypeVar, cast
 
-import numpy as np
 import torch
 from loguru import logger
-from torch import Tensor, cuda, device as Device, dtype as DType, get_rng_state, set_rng_state, stack
+from torch import Tensor, device as Device, dtype as DType, nn
 from torch.autograd import backward
-from torch.nn import Parameter
 from torch.optim import Optimizer
 from torch.optim.lr_scheduler import (
     CosineAnnealingLR,
@@ -27,73 +24,20 @@ from torch.optim.lr_scheduler import (
 from torch.utils.data import DataLoader, Dataset
 
 from refiners.fluxion import layers as fl
-from refiners.fluxion.utils import manual_seed, no_grad
+from refiners.fluxion.utils import no_grad
 from refiners.training_utils.callback import (
     Callback,
-    ClockCallback,
-    GradientNormClipping,
-    GradientValueClipping,
+    CallbackConfig,
 )
-from refiners.training_utils.config import BaseConfig, SchedulerType, TimeUnit, TimeValue
-from refiners.training_utils.dropout import DropoutCallback
-
-__all__ = ["seed_everything", "scoped_seed", "Trainer"]
-
-
-def count_learnable_parameters(parameters: Iterable[Parameter]) -> int:
-    return sum(p.numel() for p in parameters if p.requires_grad)
-
-
-def human_readable_number(number: int) -> str:
-    float_number = float(number)
-    for unit in ["", "K", "M", "G", "T", "P"]:
-        if abs(float_number) < 1000:
-            return f"{float_number:.1f}{unit}"
-        float_number /= 1000
-    return f"{float_number:.1f}E"
-
-
-def seed_everything(seed: int | None = None) -> None:
-    if seed is None:
-        seed = random.randint(0, 2**32 - 1)
-        logger.info(f"Using random seed: {seed}")
-    random.seed(a=seed)
-    np.random.seed(seed=seed)
-    manual_seed(seed=seed)
-    cuda.manual_seed_all(seed=seed)
-
-
-def scoped_seed(seed: int | Callable[..., int] | None = None) -> Callable[..., Callable[..., Any]]:
-    """
-    Decorator for setting a random seed within the scope of a function.
-
-    This decorator sets the random seed for Python's built-in `random` module,
-    `numpy`, and `torch` and `torch.cuda` at the beginning of the decorated function. After the
-    function is executed, it restores the state of the random number generators
-    to what it was before the function was called. This is useful for ensuring
-    reproducibility for specific parts of the code without affecting randomness
-    elsewhere.
-    """
-
-    def decorator(func: Callable[..., Any]) -> Callable[..., Any]:
-        @wraps(func)
-        def inner_wrapper(*args: Any, **kwargs: Any) -> Any:
-            random_state = random.getstate()
-            numpy_state = np.random.get_state()
-            torch_state = get_rng_state()
-            cuda_torch_state = cuda.get_rng_state()
-            actual_seed = seed(*args) if callable(seed) else seed
-            seed_everything(seed=actual_seed)
-            result = func(*args, **kwargs)
-            random.setstate(random_state)
-            np.random.set_state(numpy_state)
-            set_rng_state(torch_state)
-            cuda.set_rng_state(cuda_torch_state)
-            return result
-
-        return inner_wrapper
-
-    return decorator
+from refiners.training_utils.clock import ClockConfig, TrainingClock
+from refiners.training_utils.common import (
+    compute_grad_norm,
+    count_learnable_parameters,
+    human_readable_number,
+    scoped_seed,
+)
+from refiners.training_utils.config import BaseConfig, ModelConfig, SchedulerType
+from refiners.training_utils.gradient_clipping import GradientClipping, GradientClippingConfig
 
 
 class WarmupScheduler(LRScheduler):
@@ -117,135 +61,6 @@ class WarmupScheduler(LRScheduler):
             self._step_count += 1
 
 
-class TrainingClock:
-    def __init__(
-        self,
-        dataset_length: int,
-        batch_size: int,
-        training_duration: TimeValue,
-        gradient_accumulation: TimeValue,
-        evaluation_interval: TimeValue,
-        lr_scheduler_interval: TimeValue,
-    ) -> None:
-        self.dataset_length = dataset_length
-        self.batch_size = batch_size
-        self.training_duration = training_duration
-        self.gradient_accumulation = gradient_accumulation
-        self.evaluation_interval = evaluation_interval
-        self.lr_scheduler_interval = lr_scheduler_interval
-        self.num_batches_per_epoch = dataset_length // batch_size
-        self.start_time = None
-        self.end_time = None
-        self.step = 0
-        self.epoch = 0
-        self.iteration = 0
-        self.num_batches_processed = 0
-        self.num_minibatches_processed = 0
-        self.loss: Tensor | None = None
-
-    @cached_property
-    def unit_to_steps(self) -> dict[TimeUnit, int]:
-        iteration_factor = self.num_batches_per_epoch if self.gradient_accumulation["unit"] == TimeUnit.EPOCH else 1
-        return {
-            TimeUnit.STEP: 1,
-            TimeUnit.EPOCH: self.num_batches_per_epoch,
-            TimeUnit.ITERATION: self.gradient_accumulation["number"] * iteration_factor,
-        }
-
-    def convert_time_unit_to_steps(self, number: int, unit: TimeUnit) -> int:
-        return number * self.unit_to_steps[unit]
-
-    def convert_steps_to_time_unit(self, steps: int, unit: TimeUnit) -> int:
-        return steps // self.unit_to_steps[unit]
-
-    def convert_time_value(self, time_value: TimeValue, target_unit: TimeUnit) -> int:
-        number, unit = time_value["number"], time_value["unit"]
-        steps = self.convert_time_unit_to_steps(number=number, unit=unit)
-        return self.convert_steps_to_time_unit(steps=steps, unit=target_unit)
-
-    @cached_property
-    def num_epochs(self) -> int:
-        return self.convert_time_value(time_value=self.training_duration, target_unit=TimeUnit.EPOCH)
-
-    @cached_property
-    def num_iterations(self) -> int:
-        return self.convert_time_value(time_value=self.training_duration, target_unit=TimeUnit.ITERATION)
-
-    @cached_property
-    def num_steps(self) -> int:
-        return self.convert_time_value(time_value=self.training_duration, target_unit=TimeUnit.STEP)
-
-    @cached_property
-    def num_step_per_iteration(self) -> int:
-        return self.convert_time_unit_to_steps(
-            number=self.gradient_accumulation["number"], unit=self.gradient_accumulation["unit"]
-        )
-
-    @cached_property
-    def num_step_per_evaluation(self) -> int:
-        return self.convert_time_unit_to_steps(
-            number=self.evaluation_interval["number"], unit=self.evaluation_interval["unit"]
-        )
-
-    def reset(self) -> None:
-        self.start_time = None
-        self.end_time = None
-        self.step = 0
-        self.epoch = 0
-        self.iteration = 0
-        self.num_batches_processed = 0
-        self.num_minibatches_processed = 0
-
-    def start_timer(self) -> None:
-        self.start_time = time.time()
-
-    def stop_timer(self) -> None:
-        self.end_time = time.time()
-
-    @property
-    def time_elapsed(self) -> int:
-        assert self.start_time is not None, "Timer has not been started yet."
-        return int(time.time() - self.start_time)
-
-    @cached_property
-    def evaluation_interval_steps(self) -> int:
-        return self.convert_time_unit_to_steps(
-            number=self.evaluation_interval["number"], unit=self.evaluation_interval["unit"]
-        )
-
-    @cached_property
-    def lr_scheduler_interval_steps(self) -> int:
-        return self.convert_time_unit_to_steps(
-            number=self.lr_scheduler_interval["number"], unit=self.lr_scheduler_interval["unit"]
-        )
-
-    @property
-    def is_optimizer_step(self) -> bool:
-        return self.num_minibatches_processed == self.num_step_per_iteration
-
-    @property
-    def is_lr_scheduler_step(self) -> bool:
-        return self.step % self.lr_scheduler_interval_steps == 0
-
-    @property
-    def done(self) -> bool:
-        return self.step >= self.num_steps
-
-    @property
-    def is_evaluation_step(self) -> bool:
-        return self.step % self.evaluation_interval_steps == 0
-
-
-def compute_grad_norm(parameters: Iterable[Parameter]) -> float:
-    """
-    Computes the gradient norm of the parameters of a given model similar to `clip_grad_norm_` returned value.
-    """
-    gradients: list[Tensor] = [p.grad.detach() for p in parameters if p.grad is not None]
-    assert gradients, "The model has no gradients to compute the norm."
-    total_norm = stack(tensors=[gradient.norm() for gradient in gradients]).norm().item()  # type: ignore
-    return total_norm  # type: ignore
-
-
 Batch = TypeVar("Batch")
 ConfigType = TypeVar("ConfigType", bound=BaseConfig)
 
@@ -267,44 +82,91 @@ class _Dataset(Dataset[Batch]):
         return self.length
 
 
+@dataclass
+class ModelItem:
+    name: str
+    config: ModelConfig
+    model: fl.Module
+    learnable_parameters: list[nn.Parameter]
+
+
+ModelRegistry = dict[str, ModelItem]
+ModuleT = TypeVar("ModuleT", bound=fl.Module)
+
+
+def register_model():
+    def decorator(func: Callable[[Any, ModelConfig], ModuleT]) -> ModuleT:
+        @wraps(func)
+        def wrapper(self: Trainer[BaseConfig, Any], config: ModelConfig) -> fl.Module:
+            name = func.__name__
+            model = func(self, config)
+            model = model.to(self.device, dtype=self.dtype)
+            if config.requires_grad is not None:
+                model.requires_grad_(requires_grad=config.requires_grad)
+            learnable_parameters = [param for param in model.parameters() if param.requires_grad]
+            self.models[name] = ModelItem(
+                name=name, config=config, model=model, learnable_parameters=learnable_parameters
+            )
+            setattr(self, name, self.models[name].model)
+            return func(self, config)
+
+        return wrapper  # type: ignore
+
+    return decorator
+
+
+CallbackRegistry = dict[str, Callback[Any]]
+CallbackT = TypeVar("CallbackT", bound=Callback[Any])
+
+
+def register_callback():
+    def decorator(func: Callable[[Any, Any], CallbackT]) -> CallbackT:
+        @wraps(func)
+        def wrapper(self: "Trainer[BaseConfig, Any]", config: Any) -> CallbackT:
+            name = func.__name__
+            callback = func(self, config)
+            self.callbacks[name] = callback
+            setattr(self, name, callback)
+            return func(self, config)
+
+        return wrapper  # type: ignore
+
+    return decorator
+
+
 class Trainer(Generic[ConfigType, Batch], ABC):
-    def __init__(self, config: ConfigType, callbacks: list[Callback[Any]] | None = None) -> None:
+    def __init__(self, config: ConfigType) -> None:
+        self._models: ModelRegistry = {}
+        self._callbacks: CallbackRegistry = {}
         self.config = config
-        self.clock = TrainingClock(
-            dataset_length=self.dataset_length,
-            batch_size=config.training.batch_size,
-            training_duration=config.training.duration,
-            evaluation_interval=config.training.evaluation_interval,
-            gradient_accumulation=config.training.gradient_accumulation,
-            lr_scheduler_interval=config.scheduler.update_interval,
-        )
-        self.callbacks = callbacks or []
-        self.callbacks += self.default_callbacks()
+        self._load_callbacks()
         self._call_callbacks(event_name="on_init_begin")
-        self.load_models()
-        self.prepare_models()
+        self._load_models()
         self._call_callbacks(event_name="on_init_end")
 
-    def default_callbacks(self) -> list[Callback[Any]]:
-        callbacks: list[Callback[Any]] = [
-            ClockCallback(),
-            GradientValueClipping(),
-            GradientNormClipping(),
-            DropoutCallback(),
-        ]
+    @register_callback()
+    def clock(self, config: ClockConfig) -> TrainingClock:
+        return TrainingClock(
+            dataset_length=self.dataset_length,
+            batch_size=self.config.training.batch_size,
+            training_duration=self.config.training.duration,
+            evaluation_interval=self.config.training.evaluation_interval,
+            gradient_accumulation=self.config.training.gradient_accumulation,
+            lr_scheduler_interval=self.config.scheduler.update_interval,
+            verbose=config.verbose,
+        )
 
-        # look for any Callback that might be a property of the Trainer
-        for attr_name in dir(self):
-            if "__" in attr_name:
-                continue
+    @register_callback()
+    def gradient_clipping(self, config: GradientClippingConfig) -> GradientClipping:
+        return GradientClipping(config)
 
-            try:
-                attr = getattr(self, attr_name)
-            except AssertionError:
-                continue
-            if isinstance(attr, Callback):
-                callbacks.append(cast(Callback[Any], attr))
-        return callbacks
+    @property
+    def models(self) -> ModelRegistry:
+        return self._models
+
+    @property
+    def callbacks(self) -> CallbackRegistry:
+        return self._callbacks
 
     @cached_property
     def device(self) -> Device:
@@ -320,38 +182,31 @@ class Trainer(Generic[ConfigType, Batch], ABC):
         return dtype
 
     @property
-    def parameters(self) -> list[Parameter]:
-        """Returns a list of all parameters in all models"""
-        return [param for model in self.models.values() for param in model.parameters()]
-
-    @property
-    def learnable_parameters(self) -> list[Parameter]:
+    def learnable_parameters(self) -> list[nn.Parameter]:
         """Returns a list of learnable parameters in all models"""
-        return [param for model in self.models.values() for param in model.parameters() if param.requires_grad]
+        return [param for item in self.models.values() for param in item.learnable_parameters]
 
-    @property
+    @cached_property
     def optimizer_parameters(self) -> list[dict[str, Any]]:
         """
         Returns a list of `dict`-s containing the params and optimizer options for each model.
         See https://pytorch.org/docs/stable/optim.html#per-parameter-options for more details
         """
         params: list[dict[str, Any]] = []
-        for model_name, model in self.models.items():
-            model_params = [param for param in model.parameters() if param.requires_grad]
-            model_config = self.config.models[model_name]
+        for item in self.models.values():
+            config = item.config
             model_optim_conf: dict[str, Any] = {}
 
-            if model_config.learning_rate is not None:
-                model_optim_conf["lr"] = model_config.learning_rate
-            if model_config.weight_decay is not None:
-                model_optim_conf["weight_decay"] = model_config.learning_rate
-            if model_config.betas is not None:
-                model_optim_conf["betas"] = model_config.learning_rate
-            if model_config.eps is not None:
-                model_optim_conf["eps"] = model_config.learning_rate
+            if config.learning_rate is not None:
+                model_optim_conf["lr"] = config.learning_rate
+            if config.weight_decay is not None:
+                model_optim_conf["weight_decay"] = config.learning_rate
+            if config.betas is not None:
+                model_optim_conf["betas"] = config.learning_rate
+            if config.eps is not None:
+                model_optim_conf["eps"] = config.learning_rate
 
-            for param in model_params:
-                params.append({"params": param, **model_optim_conf})
+            params.append({"params": item.learnable_parameters, **model_optim_conf})
 
         return params
 
@@ -363,17 +218,12 @@ class Trainer(Generic[ConfigType, Batch], ABC):
     @property
     def gradients(self) -> list[Tensor]:
         """Returns a list of detached gradients for all learnable parameters in all models"""
-        return [
-            param.grad.detach()
-            for model in self.models.values()
-            for param in model.parameters()
-            if param.grad is not None
-        ]
+        return [param.grad.detach() for param in self.learnable_parameters if param.grad is not None]
 
     @property
     def total_gradient_norm(self) -> float:
         """Returns the total gradient norm for all learnable parameters in all models"""
-        return compute_grad_norm(parameters=self.parameters)
+        return compute_grad_norm(parameters=self.learnable_parameters)
 
     @cached_property
     def optimizer(self) -> Optimizer:
@@ -440,38 +290,6 @@ class Trainer(Generic[ConfigType, Batch], ABC):
             )
 
         return lr_scheduler
-
-    @cached_property
-    def models(self) -> dict[str, fl.Module]:
-        return self.load_models()
-
-    def set_models_to_train_mode(self) -> None:
-        for model in self.models.values():
-            model.train()
-
-    def set_models_to_eval_mode(self) -> None:
-        for model in self.models.values():
-            model.eval()
-
-    def prepare_model(self, model_name: str) -> None:
-        model = self.models[model_name]
-        if (checkpoint := self.config.models[model_name].checkpoint) is not None:
-            model.load_from_safetensors(tensors_path=checkpoint)
-        else:
-            logger.info(f"No checkpoint found. Initializing model `{model_name}` from scratch.")
-        if (requires_grad := self.config.models[model_name].requires_grad) is not None:
-            model.requires_grad_(requires_grad=requires_grad)
-        model.to(self.device)
-        model.zero_grad()
-
-    def prepare_models(self) -> None:
-        assert self.models, "No models found."
-        for model_name in self.models:
-            self.prepare_model(model_name=model_name)
-
-    @abstractmethod
-    def load_models(self) -> dict[str, fl.Module]:
-        ...
 
     @abstractmethod
     def get_item(self, index: int) -> Batch:
@@ -563,7 +381,7 @@ class Trainer(Generic[ConfigType, Batch], ABC):
     @scoped_seed(seed=get_training_seed)
     def train(self) -> None:
         """Train the model."""
-        self.set_models_to_train_mode()
+        self.set_models_to_mode("train")
         self._call_callbacks(event_name="on_train_begin")
         assert self.learnable_parameters, "There are no learnable parameters in the models."
         self.evaluate()
@@ -581,12 +399,43 @@ class Trainer(Generic[ConfigType, Batch], ABC):
     @scoped_seed(seed=get_evaluation_seed)
     def evaluate(self) -> None:
         """Evaluate the model."""
-        self.set_models_to_eval_mode()
+        self.set_models_to_mode(mode="eval")
         self._call_callbacks(event_name="on_evaluate_begin")
         self.compute_evaluation()
         self._call_callbacks(event_name="on_evaluate_end")
-        self.set_models_to_train_mode()
+        self.set_models_to_mode(mode="train")
+
+    def set_models_to_mode(self, mode: Literal["train", "eval"]) -> None:
+        for item in self.models.values():
+            if mode == "train":
+                item.model.train()
+            elif mode == "eval":
+                item.model.eval()
 
     def _call_callbacks(self, event_name: str) -> None:
-        for callback in self.callbacks:
+        for callback in self.callbacks.values():
             getattr(callback, event_name)(self)
+
+    def _load_callbacks(self) -> None:
+        for name, config in self.config:
+            if not isinstance(config, CallbackConfig):
+                continue
+            try:
+                registered_callback = getattr(self, name)
+            except AttributeError:
+                raise ValueError(
+                    f"Callback {name} is in the config but not registered in the Trainer. Create a method with the @register_callback decorator."
+                )
+            assert callable(registered_callback)
+            registered_callback(config)
+
+    def _load_models(self) -> None:
+        for name, config in self.config.models.items():
+            try:
+                registered_model = getattr(self, name)
+            except AttributeError:
+                raise ValueError(
+                    f"Model {name} is in the config but not registered in the Trainer. Create a method with the @register_model decorator."
+                )
+            assert callable(registered_model)
+            registered_model(config)
