+++ tests/training_utils/mock_config.toml
@@ -1,6 +1,11 @@
 [models.mock_model]
 requires_grad = true
 
+[clock]
+verbose = false
+
+[gradient_clipping]
+clip_grad_norm = 1.0
 
 [training]
 duration = "100:epoch"
@@ -9,7 +14,6 @@ device = "cpu"
 dtype = "float32"
 batch_size = 4
 gradient_accumulation = "4:step"
-clip_grad_norm = 1.0
 evaluation_interval = "5:epoch"
 evaluation_seed = 1
 
@@ -21,6 +25,3 @@ learning_rate = 1
 scheduler_type = "ConstantLR"
 update_interval = "1:step"
 warmup = "20:step"
-
-[dropout]
-dropout_probability = 0.0
