+++ tests/training_utils/mock_config_2_models.toml
@@ -5,12 +5,17 @@ learning_rate = 1e-5
 [models.mock_model2]
 requires_grad = true
 
+[clock]
+verbose = false
+
+[gradient_clipping]
+clip_grad_norm = 1.0
+
 [training]
 duration = "100:epoch"
 seed = 0
 batch_size = 4
 gradient_accumulation = "4:step"
-clip_grad_norm = 1.0
 evaluation_interval = "5:epoch"
 evaluation_seed = 1
 
@@ -21,6 +26,3 @@ learning_rate = 1
 [scheduler]
 scheduler_type = "ConstantLR"
 update_interval = "1:step"
-
-[dropout]
-dropout_probability = 0.0
